Add -c flag to main and refactor GetUnique
Added the -c flag to the main function.
Refactored GetUnique to call two separate functions. One with the -c
flag enabled and the other with the normal uniq behavior.

This was verified by running the original uniq unix command. Without any
input, it repeats the same input as soon as received. But when the -c
flag is enabled, it waits until it changes to return the count.